Move Game and Player into pooltool/games/__init__.py
